Refactor NativeImageClassLoaderOptions processing
Having Options defined in a JDK11 only is not compatible with
Multi-release jar standard (JEP 238 states META-INF/services is not
versioned) thus NativeImageClassLoaderOptions need to moved to base.

In case the image builder runs on classpath we need to be tolerant
towards invalid use of --add-exports -add-opens and --add-reads.  This
should be removed once we switch to unconditionally running image
builder on module-path.